requirements.txt
discord.py[voice]
requests~=2.22.0
